When calculating a perspective matrix, we should be using fov/2
This change means that most other engines will work out of the box by simply
passing the FOV value through, rather than having to half it.

Mathematically, the formula is defined as fov over 2, so we should respect that.

This also explains why setting the FOV slider to numbers greater than 90 go mental.
As technically they are using an FOV of greater than 180.